[next-react-query-tailwing] Fix imports from renaming of FileExplorerNav to RepoNavigation (#514)
* [next-react-query-tailwind] Change imports for renaming for FileExplorerNav to RepoNavigation

* [next-react-query-tailwind] Fix imports of RepoNavigation in other components
