+++ next-react-query-tailwind/src/pages/[owner]/[name]/blob/[branch]/[...path].tsx
@@ -4,7 +4,7 @@ import { withAuthRedirect } from '@lib/withAuthRedirect';
 import RepoPage from '@components/RepoPage';
 import RepoHeader from '@components/RepoHeader';
 import FileViewer from '@components/FileViewer/FileViewer.data';
-import FileExplorerNav from '@components/FileExplorerNav';
+import RepoNavigation from '@components/RepoNavigation';
 
 const RepoBranchBlob: NextPage = () => {
   const { query } = useRouter();
@@ -14,7 +14,7 @@ const RepoBranchBlob: NextPage = () => {
       <div className="max-w-screen-2xl mx-auto py-8 px-4">
         <div className="grid grid-cols-12 gap-8">
           <div className="col-span-12">
-            <FileExplorerNav />
+            <RepoNavigation />
             <FileViewer />
           </div>
         </div>
