+++ next-react-query-tailwind/src/pages/[owner]/[name]/index.tsx
@@ -5,7 +5,7 @@ import RepoPage from '@components/RepoPage';
 import FileExplorer from '@components/FileExplorer';
 import RepoHeader from '@components/RepoHeader';
 import RepoAboutWidget from '@components/RepoAboutWidget/RepoAboutWidget';
-import FileExplorerNav from '@components/FileExplorerNav';
+import RepoNavigation from '@components/RepoNavigation';
 import RepoReadMe from '@components/RepoReadMe';
 
 const RepoHome: NextPage = () => {
@@ -17,7 +17,7 @@ const RepoHome: NextPage = () => {
       <div className="max-w-screen-2xl mx-auto md:py-8 px-4">
         <div className="grid grid-cols-12 gap-8">
           <div className="col-span-12 md:col-span-7 xl:col-span-9">
-            <FileExplorerNav />
+            <RepoNavigation />
             <FileExplorer />
             <RepoReadMe />
           </div>
