+++ next-react-query-tailwind/src/pages/[owner]/[name]/tree/[branch]/[...path].tsx
@@ -4,7 +4,7 @@ import { withAuthRedirect } from '@lib/withAuthRedirect';
 import RepoPage from '@components/RepoPage';
 import FileExplorer from '@components/FileExplorer';
 import RepoHeader from '@components/RepoHeader';
-import FileExplorerNav from '@components/FileExplorerNav';
+import RepoNavigation from '@components/RepoNavigation';
 import RepoReadMe from '@components/RepoReadMe';
 
 const RepoBranchTreePath: NextPage = () => {
@@ -15,7 +15,7 @@ const RepoBranchTreePath: NextPage = () => {
       <div className="max-w-screen-2xl mx-auto py-8 px-4">
         <div className="grid grid-cols-12 gap-8">
           <div className="col-span-12">
-            <FileExplorerNav />
+            <RepoNavigation />
             <FileExplorer />
             <RepoReadMe />
           </div>
